new push

+++ C-workshop/step1/4.c
@@ -2,15 +2,15 @@
 
 void main(){
 int numbre;
-int total;
+float total;
 int i = 0;
 float moyenne;
     while(1){
         printf("Entrer un nombre:\n");
         scanf("%d",&numbre);
         if(numbre < 0){
-            moyenne = total/i;
-            printf("La moyenne est: %f",moyenne);
+            moyenne = (float)total/i;
+            printf("La moyenne est: %.2f",moyenne);
             break;
         }
         else{
